+++ src/helpers/runner/config.js
@@ -1,33 +1,22 @@
 //--------------------------------------------------------
 //-- Runner config helper
 //--------------------------------------------------------
-import { createRequire } from 'module';
-import fss               from '@absolunet/fss';
-import environment       from '../environment.js';
-import paths             from '../paths.js';
+import { createRequire } from "module";
+import fss from "@absolunet/fss";
+import environment from "../environment.js";
+import paths from "../paths.js";
 
 const require = createRequire(__filename);
 
+const JEST_GENERIC_PLUGINS = [require.resolve("@alex_neo/jest-expect-message"), require.resolve("jest-extended")];
 
-const JEST_GENERIC_PLUGINS = [
-	require.resolve('@alex_neo/jest-expect-message'),
-	require.resolve('jest-extended')
-];
-
-const JEST_PROJECT_PLUGINS = [
-	require.resolve('jest-extended')
-];
+const JEST_PROJECT_PLUGINS = [require.resolve("jest-extended")];
 
 const JEST_TRANSFORM = {
-	'\\.js$': `${paths.transformers}/babel`
+	"\\.js$": `${paths.transformers}/babel`,
 };
 
 const __ = {};
-
-
-
-
-
 
 /**
  * Configurations for Jest {@link https://jestjs.io/docs/en/configuration#projects-array-string-projectconfig project} runners.
@@ -35,7 +24,6 @@ const __ = {};
  * @hideconstructor
  */
 class RunnerHelperConfig {
-
 	/**
 	 * Jest globals variables.
 	 *
@@ -44,7 +32,6 @@ class RunnerHelperConfig {
 	get globals() {
 		return { [environment.JEST_GLOBALS_KEY]: __.globals };
 	}
-
 
 	/**
 	 * Set Jest globals variables.
@@ -55,8 +42,23 @@ class RunnerHelperConfig {
 		__.globals = globals;
 	}
 
-
 	/* eslint-disable unicorn/prevent-abbreviations */
+
+	/**
+	 * Configuration for checking files with Prettier.
+	 *
+	 * @type {object}
+	 */
+	get checkPrettier() {
+		return {
+			displayName: "Standards: Check Prettier",
+			runner: `${paths.runners}/check-prettier`,
+			rootDir: paths.project.root,
+			moduleFileExtensions: ["js", "cjs", "mjs"],
+			testMatch: ["**/*.{js,cjs,mjs}"],
+			globals: this.globals,
+		};
+	}
 
 	/**
 	 * Configuration for linting JavaScript files.
@@ -65,15 +67,14 @@ class RunnerHelperConfig {
 	 */
 	get lintJS() {
 		return {
-			displayName:          'Standards: Lint JS',
-			runner:               `${paths.runners}/lint-js`,
-			rootDir:              paths.project.root,
-			moduleFileExtensions: ['js', 'cjs', 'mjs'],
-			testMatch:            ['**/*.{js,cjs,mjs}'],
-			globals:              this.globals
+			displayName: "Standards: Lint JS",
+			runner: `${paths.runners}/lint-js`,
+			rootDir: paths.project.root,
+			moduleFileExtensions: ["js", "cjs", "mjs"],
+			testMatch: ["**/*.{js,cjs,mjs}"],
+			globals: this.globals,
 		};
 	}
-
 
 	/**
 	 * Configuration for linting JSON files.
@@ -82,14 +83,13 @@ class RunnerHelperConfig {
 	 */
 	get lintJSON() {
 		return {
-			displayName: 'Standards: Lint JSON',
-			runner:      `${paths.runners}/lint-json`,
-			rootDir:     paths.project.root,
-			testMatch:   ['**/*.json', '!**/packages/**/*'],
-			globals:     this.globals
+			displayName: "Standards: Lint JSON",
+			runner: `${paths.runners}/lint-json`,
+			rootDir: paths.project.root,
+			testMatch: ["**/*.json", "!**/packages/**/*"],
+			globals: this.globals,
 		};
 	}
-
 
 	/**
 	 * Configuration for linting YAML files.
@@ -98,15 +98,14 @@ class RunnerHelperConfig {
 	 */
 	get lintYAML() {
 		return {
-			displayName:          'Standards: Lint YAML',
-			runner:               `${paths.runners}/lint-yaml`,
-			rootDir:              paths.project.root,
-			moduleFileExtensions: ['yaml', 'yml'],
-			testMatch:            ['**/*.{yaml,yml}', '!**/packages/**/*'],
-			globals:              this.globals
+			displayName: "Standards: Lint YAML",
+			runner: `${paths.runners}/lint-yaml`,
+			rootDir: paths.project.root,
+			moduleFileExtensions: ["yaml", "yml"],
+			testMatch: ["**/*.{yaml,yml}", "!**/packages/**/*"],
+			globals: this.globals,
 		};
 	}
-
 
 	/**
 	 * Configuration for linting Bash files.
@@ -115,15 +114,14 @@ class RunnerHelperConfig {
 	 */
 	get lintBash() {
 		return {
-			displayName:          'Standards: Lint Bash',
-			runner:               `${paths.runners}/lint-bash`,
-			rootDir:              paths.project.root,
-			moduleFileExtensions: ['sh'],
-			testMatch:            ['**/*.sh', '!**/packages/**/*'],
-			globals:              this.globals
+			displayName: "Standards: Lint Bash",
+			runner: `${paths.runners}/lint-bash`,
+			rootDir: paths.project.root,
+			moduleFileExtensions: ["sh"],
+			testMatch: ["**/*.sh", "!**/packages/**/*"],
+			globals: this.globals,
 		};
 	}
-
 
 	/**
 	 * Configuration for linting SCSS files.
@@ -132,15 +130,14 @@ class RunnerHelperConfig {
 	 */
 	get lintSCSS() {
 		return {
-			displayName:          'Standards: Lint SCSS',
-			runner:               `${paths.runners}/lint-scss`,
-			rootDir:              paths.project.root,
-			moduleFileExtensions: ['scss'],
-			testMatch:            ['**/*.scss'],
-			globals:              this.globals
+			displayName: "Standards: Lint SCSS",
+			runner: `${paths.runners}/lint-scss`,
+			rootDir: paths.project.root,
+			moduleFileExtensions: ["scss"],
+			testMatch: ["**/*.scss"],
+			globals: this.globals,
 		};
 	}
-
 
 	/**
 	 * Configuration for linting files styling.
@@ -149,24 +146,23 @@ class RunnerHelperConfig {
 	 * @returns {object} Configuration.
 	 */
 	lintFileStyles(repositoryType) {
-		const prefix = repositoryType === environment.REPOSITORY_TYPE.subPackage ? `/../..` : '';
+		const prefix = repositoryType === environment.REPOSITORY_TYPE.subPackage ? `/../..` : "";
 
-		const rawConfig   = fss.readFile(`${paths.project.root}${prefix}/.editorconfig`, 'utf8');
-		const rawPatterns = [...rawConfig.matchAll(/^\[(?<pattern>.+)\]$/gum)];
-		const patterns    = rawPatterns.map((item) => {
+		const rawConfig = fss.readFile(`${paths.project.root}${prefix}/.editorconfig`, "utf8");
+		const rawPatterns = [...rawConfig.matchAll(/^\[(?<pattern>.+)\]$/gmu)];
+		const patterns = rawPatterns.map((item) => {
 			return `**/${item[1]}`;
 		});
 
 		return {
-			displayName:          'Standards: Lint file styles',
-			runner:               `${paths.runners}/lint-file-styles`,
-			rootDir:              paths.project.root,
-			moduleFileExtensions: ['*'],
-			testMatch:            [...patterns, '!**/*.{js,scss}', '!**/packages/**/*'],
-			globals:              this.globals
+			displayName: "Standards: Lint file styles",
+			runner: `${paths.runners}/lint-file-styles`,
+			rootDir: paths.project.root,
+			moduleFileExtensions: ["*"],
+			testMatch: [...patterns, "!**/*.{js,scss}", "!**/packages/**/*"],
+			globals: this.globals,
 		};
 	}
-
 
 	/**
 	 * Configuration for validating a repository.
@@ -175,14 +171,13 @@ class RunnerHelperConfig {
 	 */
 	get genericTests() {
 		return {
-			displayName:        'Standards: Generic tests',
-			rootDir:            `${paths.project.test}/generic`,
+			displayName: "Standards: Generic tests",
+			rootDir: `${paths.project.test}/generic`,
 			setupFilesAfterEnv: JEST_GENERIC_PLUGINS,
-			transform:          JEST_TRANSFORM,
-			globals:            this.globals
+			transform: JEST_TRANSFORM,
+			globals: this.globals,
 		};
 	}
-
 
 	/**
 	 * Configuration for running a project's custom standards tests.
@@ -191,13 +186,12 @@ class RunnerHelperConfig {
 	 */
 	get projectStandardsTests() {
 		return {
-			displayName:        'Standards: Project tests',
-			rootDir:            `${paths.project.test}/standards`,
+			displayName: "Standards: Project tests",
+			rootDir: `${paths.project.test}/standards`,
 			setupFilesAfterEnv: JEST_PROJECT_PLUGINS,
-			transform:          JEST_TRANSFORM
+			transform: JEST_TRANSFORM,
 		};
 	}
-
 
 	/**
 	 * Configuration for running a project's custom unit tests.
@@ -206,13 +200,12 @@ class RunnerHelperConfig {
 	 */
 	get projectUnitTests() {
 		return {
-			displayName:        'Unit: Project tests',
-			rootDir:            `${paths.project.test}/unit`,
+			displayName: "Unit: Project tests",
+			rootDir: `${paths.project.test}/unit`,
 			setupFilesAfterEnv: JEST_PROJECT_PLUGINS,
-			transform:          JEST_TRANSFORM
+			transform: JEST_TRANSFORM,
 		};
 	}
-
 
 	/**
 	 * Configuration for running a project's custom feature tests.
@@ -221,13 +214,12 @@ class RunnerHelperConfig {
 	 */
 	get projectFeatureTests() {
 		return {
-			displayName:        'Feature: Project tests',
-			rootDir:            `${paths.project.test}/feature`,
+			displayName: "Feature: Project tests",
+			rootDir: `${paths.project.test}/feature`,
 			setupFilesAfterEnv: JEST_PROJECT_PLUGINS,
-			transform:          JEST_TRANSFORM
+			transform: JEST_TRANSFORM,
 		};
 	}
-
 
 	/**
 	 * Configuration for running a project's custom integration tests.
@@ -236,13 +228,12 @@ class RunnerHelperConfig {
 	 */
 	get projectIntegrationTests() {
 		return {
-			displayName:        'Integration: Project tests',
-			rootDir:            `${paths.project.test}/integration`,
+			displayName: "Integration: Project tests",
+			rootDir: `${paths.project.test}/integration`,
 			setupFilesAfterEnv: JEST_PROJECT_PLUGINS,
-			transform:          JEST_TRANSFORM
+			transform: JEST_TRANSFORM,
 		};
 	}
-
 
 	/**
 	 * Configuration for running a project's custom end-to-end tests.
@@ -251,16 +242,14 @@ class RunnerHelperConfig {
 	 */
 	get projectEndtoendTests() {
 		return {
-			displayName:        'End-to-end: Project tests',
-			rootDir:            `${paths.project.test}/endtoend`,
+			displayName: "End-to-end: Project tests",
+			rootDir: `${paths.project.test}/endtoend`,
 			setupFilesAfterEnv: JEST_PROJECT_PLUGINS,
-			transform:          JEST_TRANSFORM
+			transform: JEST_TRANSFORM,
 		};
 	}
 
 	/* eslint-enable unicorn/prevent-abbreviations */
-
 }
-
 
 export default new RunnerHelperConfig();