+++ src/runners/lint-scss/run.js
@@ -1,20 +1,20 @@
 //--------------------------------------------------------
 //-- Lint SCSS runner - Run
 //--------------------------------------------------------
-import { pass, fail, skip } from 'create-jest-runner';
-import stylelint            from 'stylelint';
-import runner               from '../../helpers/runner/index.js';
-
+import { pass, fail, skip } from "create-jest-runner";
+import stylelint from "stylelint";
+import runner from "../../helpers/runner/index.js";
 
 const lintSCSSRunner = ({ testPath }) => {
-	const testResult = runner.initTestResult({ testPath, title: 'stylelint' });
+	const testResult = runner.initTestResult({ testPath, title: "stylelint" });
 
-	return stylelint.lint({
-		files:           testPath,
-		syntax:          'scss',
-		formatter:       'string',
-		allowEmptyInput: true
-	})
+	return stylelint
+		.lint({
+			files: testPath,
+			syntax: "scss",
+			formatter: "string",
+			allowEmptyInput: true,
+		})
 		.then((data) => {
 			const [results] = data.results;
 
@@ -23,20 +23,17 @@ const lintSCSSRunner = ({ testPath }) => {
 			}
 
 			if (results.warnings.length > 0 || results.deprecations.length > 0 || results.invalidOptionWarnings.length > 0) {
-				const rawOutput = data.output.split('\n');
+				const rawOutput = data.output.split("\n");
 				rawOutput.splice(0, 2);
 
-				return fail(testResult(rawOutput.join('\n')));
+				return fail(testResult(rawOutput.join("\n")));
 			}
 
 			return pass(testResult());
 		})
 		.catch((error) => {
 			return fail(testResult(runner.formatError(`${error.stack}\n`)));
-		})
-	;
-
+		});
 };
-
 
 export default lintSCSSRunner;