+++ src/runners/lint-yaml/index.js
@@ -1,6 +1,6 @@
 //--------------------------------------------------------
 //-- Lint YAML runner
 //--------------------------------------------------------
-import runner from '../../helpers/runner/index.js';
+import runner from "../../helpers/runner/index.js";
 
 export default runner.create(__dirname);