+++ src/tests/repository/arborescence.js
@@ -1,12 +1,10 @@
 //--------------------------------------------------------
 //-- Repository - Arborescence tests
 //--------------------------------------------------------
-import arborescence from '../../helpers/arborescence.js';
-
+import arborescence from "../../helpers/arborescence.js";
 
 const arborescenceTests = (options) => {
 	arborescence.validate(options);
 };
-
 
 export default arborescenceTests;