+++ CHANGELOG.md
@@ -13,6 +13,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 
+## [5.0.0-beta.2] - 2021-09-16
+### Added
+- Validate Prettier formatting and configuration
+
+### Changed
+- ESLint config update to 3.0.0-beta.3
+- Maintenance update
+
+### Fixed
+- Fix selection process for file matrix 
+- Allow multi-packages to be CommonJS
+
+### Removed
+- Stop testing `package-lock.json` since it is a generated file 
+
+
+
 ## [5.0.0-beta.1] - 2021-09-10
 ### Added
 - Validate ESM and CommonJS packages
@@ -547,7 +564,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 
 
-[Unreleased]:   https://github.com/absolunet/node-tester/compare/5.0.0-beta.1...HEAD
+[Unreleased]:   https://github.com/absolunet/node-tester/compare/5.0.0-beta.2...HEAD
+[5.0.0-beta.2]: https://github.com/absolunet/node-tester/compare/5.0.0-beta.1...5.0.0-beta.2
 [5.0.0-beta.1]: https://github.com/absolunet/node-tester/compare/4.2.0...5.0.0-beta.1
 [4.2.0]:        https://github.com/absolunet/node-tester/compare/4.1.4...4.2.0
 [4.1.4]:        https://github.com/absolunet/node-tester/compare/4.1.3...4.1.4